Move get_iap_garage function to common.utils
