+++ iap/utils.py
@@ -1,16 +1,10 @@
 import datetime
-import os
-from typing import List
 
-from sqlalchemy import func, Date, cast, select, distinct
+from sqlalchemy import func, Date, cast
 
 from common import logger
-from common._crypto import Account
 from common.enums import ReceiptStatus
-from common.models.garage import GarageItemStatus
-from common.models.product import FungibleItemProduct
 from common.models.receipt import Receipt
-from common.utils.aws import fetch_kms_key_id
 
 
 def get_purchase_count(sess, agent_addr: str, product_id: int, hour_limit: int) -> int:
@@ -37,21 +31,3 @@ def get_purchase_count(sess, agent_addr: str, product_id: int, hour_limit: int) 
         f"Agent {agent_addr} purchased product {product_id} {purchase_count} times in {hour_limit} hours from {start}"
     )
     return purchase_count
-
-
-def get_iap_garage(sess) -> List[GarageItemStatus]:
-    """
-    Get NCG balance and fungible item count of IAP address.
-    :return:
-    """
-    stage = os.environ.get("STAGE", "development")
-    region_name = os.environ.get("REGION_NAME", "us-east-2")
-    account = Account(fetch_kms_key_id(stage, region_name))
-
-    fungible_id_list = sess.scalars(select(distinct(FungibleItemProduct.fungible_item_id))).fetchall()
-    return sess.scalars(
-        select(GarageItemStatus).where(
-            GarageItemStatus.address == account.address,
-            GarageItemStatus.fungible_id.in_(fungible_id_list)
-        )
-    )
